+++ CHANGELOG.md
@@ -45,6 +45,7 @@ The change log for older releases (before 5.9.0) can be found in CHANGELOG.old.
 
 ### Fixed
  - Fix crash when trying to unset the untagged tag via the settings dialog (#477529)
+ - Fix crash when renaming or adding a category and using it as untagged category (#477530)
 
 ### Removed